pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kalshi-alpha"
version = "0.1.0"
description = "Monorepo for Kalshi ladder pricing, scanning, and research."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Kalshi Alpha Team" },
]
license = { text = "MIT" }
dependencies = [
  "requests>=2.32.3",
  "pydantic>=2.7.0",
  "polars>=1.6.0",
  "duckdb>=0.10.2",
  "python-dateutil>=2.9.0.post0",
  "pyyaml>=6.0.1",
  "cachetools>=4.2,<5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "hypothesis>=6.99.0",
  "ruff>=0.5.6",
  "mypy>=1.10.0",
  "types-requests>=2.32.0.20240914",
  "types-python-dateutil>=2.9.0.20240821",
  "types-pyyaml>=6.0.12.20240813",
  "pre-commit>=3.7.1",
]

[project.scripts]
kalshi-scan = "kalshi_alpha.exec.runners.scan_ladders:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["kalshi_alpha*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_equality = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["duckdb", "polars", "cachetools"]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S", "DTZ", "TID", "ERA", "PL", "ANN"]
ignore = ["S101", "PLR2004"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR2004"]

[tool.ruff.lint.isort]
known-first-party = ["kalshi_alpha"]

[tool.ruff.format]
quote-style = "double"

[tool.coverage.run]
branch = true
omit = ["tests/*", "notebooks/*"]
